Fix unexpected char writing with limitation
The pixel_clk-synchronized writing will asynchronizely stall and recover the pipeline, which introduces
the following problem:

If we don't immediately stall the pipeline when accessing text memory using a combinational logic on address,
it will go on as other signals are synchronized with pixel_clk which may raise after a while.
However, the combinational logic, a.k.a vga_stall, is so beyond control that a consecutive accessing of text memory
is unresolvable! The vga_stall_cnt needs an inactive vga_stall to reset itself which cannot be feasible under such
a circumstance as the vga_stall is always high. Therefore it ignores following writing to text memory.

We can try reset vga_stall_cnt early, but this will result in a endless mem_stall with just a small
interval for pipeline to continue, if the following instructions all write to text memory, and will
introduce race as vga_wen is always high.

This situation has two patterns:

1. sb ch1, addr
   sb ch2, addr

2. sb ch, addr1
   sb ch, addr2

The 1st one is stupid. We should throw away the compiler or the asm coder.
The 2nd one will occur typically with the following code:

  char s[] = "aaa";

which is common but can be avoided.
How to trade off between the simple c and the simple verilog needs further discussion.