Fix silly misprediction in bpu
Originally, when running into a trap or endless loop, the mem-pc
blinked and was difficult to figure out which trap the cpu is in.

As absolutely branching to itself is a fixed pattern, the bpu
should at last give the right prediction, but it failed. The
problem is that I used a wrong tag to index the BTB when storing
the branch target. Consequently, the right branch target was in
another (and wrong) BTB entry while the prediction logic read from
the right entry, but ether it was invalid or its value was wrong.

Jmp won't write BTB, so we need to endure the blinking jmp trap...