Remove delay logic for loader
After testing, the block ram with no primitive register output
can response in the same clock when request arives.

But, a strange phenomenon is that as the timing graph describes, we
should wait another ui_clk to sample the addr, data and wen, but
a reversed clock like register files' results in wrong execution
(finally, instruction fetch fails). The current design sharing the
same clock with pipeline is anti-intuitive, but it works fine...