Fix sigh-unsign mismatch in hand-written divison
The divison takes place in a unsigned context.
The previous signed quotient and remain compared to unsigned value
will result in unexpected early exit of the logic.
For example, when we want to print a signed value in a unsigned
manner, it will end with an unexpected '-' and no digits left.