Seperate object files to avoid conflict
Before, when we build NEMU inside ram_init_gen, the object files
are generated in the original directory, which are mips elf files.
If we want to build NEMU with both of the NPC version and the host
version, we might fall into endless switching between make clean
and cd.

P.S. the double check macro in loader.c is deprecated as we don't
need it.